UBUNTU: SAUCE: driver: cpufreq: put offline cpu cores to fmin
BugLink: https://bugs.launchpad.net/bugs/2012939

Offline CPU's in a cluster also participate in the final frequency
determination for that cluster as the max of currently requested
freq to all CPU's of the cluster is selected. So, an offline CPU
with stale high freq request can override the current low freq
request to the online CPU's. To fix this, set the frequency of
offline CPU to Fmin in hotplug off notifier so it can't override
the freq request to the online CPU's in the same cluster. Again
during system resume, MCE updates that Fmin for offline cores to
the default configured value as the MCE doesn't differentiate
between normal boot and system resume. This is resulting in high
power consumption. So, updating the freq of offline cores to Fmin
during system resume also.

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Pohsun Su <pohsuns@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Ian May <ian.may@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Brad Figg <bfigg@nvidia.com>