ANDROID: KVM: arm64: Remove the non-consuming hyp trace interface
Following ML discussion, it seems very much likely the non-consuming
upstream version for hyp tracing ("trace" files) will be delayed.

As we want to stick to upstream as much as possible for the next Android
releases, let's remove this interface, not used by our tools before
anyone relies on it.

Bug: 249050813
Link: https://lore.kernel.org/linux-trace-kernel/20230329070353.1e1b443b@gandalf.local.home/
Change-Id: I64b032512b85ed095bb402dc2997101777cb539b
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>